requirements.txt
# Versions as of 2015-01-15

# Core
django==1.8.3
djangorestframework==3.1.3
django-appconf==1.0.1
django-filter==0.9.1
django-ordered-model==0.4.2
django-modeltranslation==0.10
django-compressor==1.5
Pillow==2.9.0
reportlab==3.2.0

# Sample data
factory_boy==2.5.2

# Goodies
django-extensions==1.5.5
